Make the parsing faster in Mojolicious environment
For some reason - most likely some special pragma, when parsing explain with
very long lines, and "QUERY PLAN" header (which in turn contains a lot of
spaces, parsing took very long time.

Simple change from:
( \s* (->)? \s* )
to
( \s* -> \s* | \s* )
made the parsing time fast again, even in this environment.

I was not able to replicate the problem on standalone test case, so I don't
know exactly what is wrong with the code when run inside Mojolicious, but
since Mojolicious is the base of http://explain.depesz.com/ - it's important
problem, so it has to be fixed.